requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
flask==3.1.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==3.0.2
python-decouple==3.8
requests==2.32.3
urllib3==2.2.3
werkzeug==3.1.3
